Add files via upload
The CNN classification model uses pre-trained MobileNetv2 and is developed in Python using Keras and Tensor flow in the backend. The model is developed and tested with different network architectures by varying the type of layers used to train the network including convolution layers, Dropout layers, pooling layers and Dense layers. The model uses Transfer Learning techniques for early convergence. It was tested and trained on the dataset collected and derived from Kaggle - Face mask Dataset which is a benchmark public database for machine learning. The images are pre-processed using data augmentation technique. The model achieved 99.39% accuracy and is accurately detecting images with mask and without mask.